+++ business/src/useCases/createACaseProxy.js
@@ -1,12 +1,11 @@
 const axios = require('axios');
 
 module.exports = async ({ userId, documents, applicationContext }) => {
-  const baseUrl = applicationContext.getBaseUrl();
   const headers = {
     Authorization: `Bearer ${userId}`,
   };
   const response = await axios.post(
-    `${baseUrl}/cases`,
+    `${applicationContext.getBaseUrl()}/cases`,
     {
       documents,
     },
