+++ web-client/src/main/actions/clearAlerts.js
@@ -1,0 +1,6 @@
+import { state } from 'cerebral';
+
+export default ({ store }) => {
+  store.set(state.alertError, null);
+  store.set(state.alertSuccess, null);
+};
