+++ web-client/src/main/actions/clearLoginForm.js
@@ -1,0 +1,7 @@
+import { state } from 'cerebral';
+
+export default ({ store }) => {
+  store.set(state.form, {
+    name: '',
+  });
+};
