+++ web-client/src/main/actions/clearPetition.js
@@ -1,0 +1,10 @@
+import { state } from 'cerebral';
+
+export default ({ store }) => {
+  store.set(state.petition, {
+    petitionFile: null,
+    requestForPlaceOfTrial: null,
+    statementOfTaxpayerIdentificationNumber: null,
+    uploadsFinished: 0,
+  });
+};
