+++ web-client/src/main/actions/getCaseDetail.js
@@ -1,0 +1,10 @@
+import { state } from 'cerebral';
+
+export default async ({ applicationContext, get, props, useCases }) => {
+  const caseDetail = await useCases.getCaseDetail(
+    applicationContext,
+    props.caseId,
+    get(state.user.userId),
+  );
+  return { caseDetail };
+};
