+++ web-client/src/main/actions/getCreateCaseAlertSuccess.js
@@ -1,0 +1,8 @@
+export default () => {
+  return {
+    alertSuccess: {
+      title: 'Your files were uploaded successfully.',
+      message: 'Your case has now been created.',
+    },
+  };
+};
