+++ web-client/src/main/actions/getPetitionsClerkCaseList.js
@@ -1,0 +1,18 @@
+import { state } from 'cerebral';
+
+export default async ({ applicationContext, get, path, useCases }) => {
+  try {
+    const caseList = await useCases.getPetitionsClerkCaseList(
+      applicationContext,
+      get(state.user.userId),
+    );
+    return path.success({ caseList });
+  } catch (e) {
+    return path.error({
+      alertError: {
+        title: 'Cases not found',
+        message: 'There was a problem getting the cases',
+      },
+    });
+  }
+};
