+++ web-client/src/main/actions/getUserRole.js
@@ -1,0 +1,6 @@
+import { state } from 'cerebral';
+
+export default ({ get, path }) => {
+  const user = get(state.user);
+  return path[user.role]();
+};
