+++ web-client/src/main/actions/navigateToDashboard.js
@@ -1,0 +1,3 @@
+export default ({ router }) => {
+  router.route('/');
+};
