+++ web-client/src/main/actions/setAlertError.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ props, store }) => {
+  store.set(state.alertError, props.alertError);
+};
