+++ web-client/src/main/actions/setAlertSuccess.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ props, store }) => {
+  store.set(state.alertSuccess, props.alertSuccess);
+};
