+++ web-client/src/main/actions/setBaseUrl.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ store, applicationContext }) => {
+  store.set(state.baseUrl, applicationContext.getBaseUrl());
+};
