+++ web-client/src/main/actions/setCaseDetail.js
@@ -1,0 +1,6 @@
+import { state } from 'cerebral';
+
+export default ({ store, props }) => {
+  store.set(state.caseDetail, props.caseDetail);
+  return;
+};
