+++ web-client/src/main/actions/setCaseList.js
@@ -1,0 +1,6 @@
+import { state } from 'cerebral';
+
+export default ({ store, props }) => {
+  store.set(state.cases, props.caseList);
+  return;
+};
