+++ web-client/src/main/actions/setFormSubmitting.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ store }) => {
+  store.set(state.submitting, true);
+};
