+++ web-client/src/main/actions/setUser.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ store, props }) => {
+  store.set(state.user, props.user);
+};
