+++ web-client/src/main/actions/unsetFormSubmitting.js
@@ -1,0 +1,5 @@
+import { state } from 'cerebral';
+
+export default ({ store }) => {
+  store.set(state.submitting, false);
+};
